+++ com.unity.shadergraph/CHANGELOG.md
@@ -51,6 +51,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Newly created properties and keywords will no longer use obfuscated GUID-based reference names in the shader code [1300484]
 - Fixed ParallaxMapping node compile issue on GLES2
 - Fixed a selection bug with block nodes after changing tabs [1312222]
+- Fixed some shader graph compiler errors not being logged [1304162].
 
 ## [10.3.0] - 2020-11-03